Добавляет webpack, настройки по проекту, и разделяет код на модули
