+++ src/components/filmStatistics.js
@@ -1,0 +1,7 @@
+const createFilmStatistics = () => {
+  return (`
+    <p>0 movies inside</p>
+  `);
+};
+
+export default createFilmStatistics;
