+++ src/components/showButton.js
@@ -1,0 +1,7 @@
+const createShowButtonTemplate = () => {
+  return (`
+    <button class="films-list__show-more">Show more</button>
+  `);
+};
+
+export default createShowButtonTemplate;
